requirements.txt
django==3.1.6
psycopg2==2.8.6
djangorestframework==3.12.4
django-widget-tweaks==1.4.8
django-crispy-forms==1.11.1
numpy==1.20.1
xlsxwriter==1.2.7
xlrd==1.2.0
pandas==1.2.2